added all notifications

+++ src/main/java/DomainLayer/DomainServices/ToNotify.java
@@ -3,11 +3,17 @@ package DomainLayer.DomainServices;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import DomainLayer.IUserRepository;
+import DomainLayer.Roles.RegisteredUser;
 import InfrastructureLayer.NotificationRepository;
 import DomainLayer.IToken;
 
 
+import InfrastructureLayer.UserRepository;
+import ServiceLayer.OwnerManagerService;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.vaadin.flow.component.notification.Notification;
 import utils.Notifications;
 
 
@@ -16,16 +22,18 @@ public class ToNotify {
     private IToken tokenService;
     private NotificationWebSocketHandler notificationWebSocketHandler;
     private final ObjectMapper mapper = new ObjectMapper();
+    private UserRepository userRepository;
 
-    public ToNotify(NotificationRepository notificationRepo, IToken tokenService, NotificationWebSocketHandler notificationWebSocketHandler) {
+    public ToNotify(NotificationRepository notificationRepo, IToken tokenService, NotificationWebSocketHandler notificationWebSocketHandler, UserRepository userRepository) {
         this.tokenService = tokenService;
         this.notificationRepo = notificationRepo;
         this.notificationWebSocketHandler = notificationWebSocketHandler;
+        this.userRepository = userRepository;
     }
 
     public List<Notifications> getUserNotifications(String token) {
         String receiverUsername = tokenService.extractUsername(token);
-        List<Notifications> notifications = notificationRepo.findByUserID(receiverUsername);
+        List<Notifications> notifications = notificationRepo.getAll(); //.findByUserID(receiverUsername);
         ArrayList<Notifications> messages = new ArrayList<>();
         for (Notifications notification : notifications) {
             messages.add(notification);
@@ -54,7 +62,7 @@ public class ToNotify {
             if(tokenService.getToken(userId).equals("")) {
                 notificationRepo.save(notification);
             } else {
-                notificationWebSocketHandler.sendNotificationToClient(userId, notification.getMessage());
+                notificationWebSocketHandler.sendNotificationToClient(userId, notification.getUserId());
             }
         } catch (Exception e) {
             throw new RuntimeException("Failed to serialize notification", e);
@@ -66,13 +74,22 @@ public class ToNotify {
             List<Notifications> notifications = getUserNotifications(token);
             for (Notifications notification : notifications) {
                 String username = tokenService.extractUsername(token);
-                notificationWebSocketHandler.sendNotificationToClient(username, notification.getMessage());
-                notificationRepo.delete(notification);
-
+                RegisteredUser user = null;
+                try {
+                    user = userRepository.getById(username);
+                } catch (Exception e) {
+                }
+                List<String> managedStores = user.getManagedStores();
+                List<String> managedStoresfg = managedStores;
+                if (notification.getStoreId().equals(username) && (notification.getMessage().equals("") || managedStoresfg.contains(notification.getMessage()))) {
+                    notificationWebSocketHandler.sendNotificationToClient(username, notification.getUserId());
+                    notificationRepo.delete(notification);
+                }
             }
         } catch (Exception e) {
 
         }
     }
+
 
 }
